Add files via upload
Text Embedding

Today, social networks have a wide range of applications. Its first use is entertainment and leisure. But from another perspective, social networks can be used to find behavioral patterns. For example, by analyzing the opinions of social network users, we can find the weaknesses of our business. "Gender" is one of the parameters that influence user behavior. When faced with an issue, women will mostly react in one way and men in another! 

During the following codes, we intend to predict their gender using the information that Datak has provided us with from Twitter and Instagram users.