Add files via upload
NLP

Real Persian web data, filtered and collected by the Yektant platform, has been provided to us. The goal is to build a machine learning model that can predict the topic category of a document based on the texts in a link, such as title, description, text content, etc. For example, if we have a link from a news site titled "Kayhan Kalhor Receives World Music Man of the Year Award", our model should predict that this article is related to the topic of "music". Of course, in this example, we only mentioned the title feature, while the description text or text content can also be used.