Add files via upload
NLP

You have noticed that on websites, when you are typing into a text field, a list of similar texts is usually suggested to you. However, such suggested lists are usually very simple and only suggest phrases that start with the user's typed string so that the user does not need to complete their typing and gets the result faster. We want to build a suggester that is smarter and more flexible. For example, maybe the user typed the phrase "Babel" but initially made a spelling mistake and meant "Zabol". Or maybe, as you must have done and got on your nerves, they forgot to set their keyboard to Persian and typed a phrase like "fhfg". Also, why not consider the English names of cities and when someone types, for example, "Zahedan", we do not realize that they meant "Zahedan"? However, these are just a few of the possible patterns that will lead to improvements in the system's suggestions. We can use the data collected from what users typed step by step and what they ultimately selected to design a model that, given the user's input, suggests a list of the most likely choices.